new factory for creating aggregations
