--- source/Application/Transactions/Aggregations/IGridAreaLookup.cs
@@ -1,31 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System.Threading.Tasks;
-using Domain.Actors;
-
-namespace Application.Transactions.Aggregations;
-
-/// <summary>
-/// Service looking up grid area details
-/// </summary>
-public interface IGridAreaLookup
-{
-    /// <summary>
-    /// Finds the actor number of the grid operator of the specified grid area
-    /// </summary>
-    /// <param name="gridAreaCode"></param>
-    /// <returns><see cref="ActorNumber"/></returns>
-    Task<ActorNumber> GetGridOperatorForAsync(string gridAreaCode);
-}
