+++ source/IntegrationTests/TestBase.cs
@@ -187,7 +187,6 @@ namespace IntegrationTests
 
             _services.AddTransient<InboxEventsProcessor>();
             _services.AddTransient<AggregatedTimeSeriesRequestAcceptedEventMapper>();
-            _services.AddTransient<IGridAreaLookup, GridAreaLookup>();
             _services.AddTransient<INotificationHandler<TestNotification>>(_ => NotificationHandlerSpy);
             _services.AddTransient<INotificationHandler<AggregatedTimeSeriesRequestWasAccepted>>(_ => TestAggregatedTimeSeriesRequestAcceptedHandlerSpy);
             _services.AddTransient<INotificationHandler<IntegrationTests.Infrastructure.InboxEvents.TestNotification>>(
